requirements.txt
psycopg2-binary
Jinja2
mistune
pyrsistent
six
Werkzeug
zipp
ruamel.yaml
APIFlask==2.4.0
Flask-Cors
Flask-Limiter
kubernetes
gunicorn
gitpython 
boto3 
pyyaml
dbt-bigquery==1.9.2
dbt-snowflake==1.9.4
dbt-redshift==1.9.5
dbt-fabric==1.9.6
minio